Added pages for client functions and user functions.
There are now php pages for user defined functions that will be used repeatedly. One is for client related functions and the other is for user related functions.

I also changed the accounts page so that it loads user details from the database. Currently, they are not being displayed however. Working on fixing that.